API通信のファイルを分ける
