+++ app/src/main/java/com/android/example/tanmen/Controller/Fragment/MainFragment.kt
@@ -10,6 +10,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.setFragmentResult
 import androidx.fragment.app.setFragmentResultListener
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
 import com.android.example.tanmen.Model.Shop
 import com.android.example.tanmen.R
 import com.android.example.tanmen.View.ShopListAdapter
@@ -18,6 +20,7 @@ import com.android.example.tanmen.databinding.FragmentMainBinding
 class MainFragment : Fragment() {
     private lateinit var binding: FragmentMainBinding
     var adapter: ShopListAdapter? = null
+    var shopList = mutableListOf<Shop>()
 
     companion object {
         const val REQ_KEY: String = "shop"
@@ -33,15 +36,15 @@ class MainFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         binding = FragmentMainBinding.inflate(inflater, container, false)
+        val recyclerView = binding.shopList
+        val layoutManager = LinearLayoutManager(recyclerView.context)
+        adapter = ShopListAdapter(shopList)
+        recyclerView.layoutManager = layoutManager
+        recyclerView.adapter = adapter
+
         setFragmentResultListener(REQ_KEY) { _, bundle ->
             val shopLists = bundle.getSerializable(ARG_SHOP) as MutableList<Shop>
-
-            val recyclerView = binding.shopList
-            val layoutManager = LinearLayoutManager(recyclerView.context)
-            adapter = ShopListAdapter(shopLists)
-            recyclerView.layoutManager = layoutManager
-            recyclerView.adapter = adapter
-
+            changeShopList(shopLists)
         }
 
         adapter?.setOnShopCellClickListener(
@@ -63,6 +66,12 @@ class MainFragment : Fragment() {
             }
         )
         return binding.root
+    }
+
+    private fun changeShopList(shopLists: MutableList<Shop>) {
+        val _adapter = binding.shopList.adapter as ShopListAdapter
+        _adapter.shopList = shopLists
+        _adapter.notifyDataSetChanged()
     }
 
 }
